splits tests by operations and
